requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.18
pymupdf==1.25.3
google-generativeai==0.8.3
chromadb==0.6.3
python-dotenv==1.0.1
python-docx==1.1.2
pydantic==2.10.4
